chore: Player 정보 Room에 정의
